update hives pagination
